Reformat carbonblack transmission module
